+++ index.html
@@ -3,7 +3,6 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <script src="scripts/new.js" defer></script>
   <link rel="stylesheet" href="styles/styles.css">
   <title>Awesome Books</title>
 </head>
@@ -24,8 +23,11 @@
     </div>
   </div>
 
+  <div>
+    <p class="time"></p>
+   </div>
   <!-- Book List Section -->
-
+   
   <div class="book-list-section">
     <div class="book-list-container">
       <h1 class="awesome-books">All Awesome Boooks</h1>
@@ -75,6 +77,6 @@
       </p>
     </div>
   </footer>
-  
+  <script src="/index.js" type="module"></script>
 </body>
 </html>
